# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express Server API 文档",
    "version": "0.0.1",
    "description": "床前明月光，疑是地上霜。举头望明月，低头思故乡。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1/api"
    }
  ],
  "paths": {
...
    "/blog/articles": {
      "get": {
        "summary": "获取文章列表",
        "description": "分页获取文章列表，支持标签和分类过滤",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取文章列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "articles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "title": {
                            "type": "string"
                          },
                          "content": {
                            "type": "string"
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "name": {
                                  "type": "string"
                                }
                              }
                            }
                          },
                          "category": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "获取文章列表失败"
          }
        }
      },
      "post": {
        "summary": "创建新文章",
        "description": "创建一篇新的博客文章",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content", "tags", "category"],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "文章标题"
                  },
                  "content": {
                    "type": "string",
                    "description": "文章内容"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "文章标签ID列表"
                  },
                  "category": {
                    "type": "string",
                    "description": "文章分类ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "文章创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "article": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "创建文章失败"
          }
        }
      }
    },
    "/blog/articles/{id}": {
      "get": {
        "summary": "获取文章详情",
        "description": "根据文章ID获取文章详细信息",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "文章唯一标识符"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取文章详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "article": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        },
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "_id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "category": {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            }
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "文章未找到"
          },
          "500": {
            "description": "获取文章详情失败"
          }
        }
      },
      "put": {
        "summary": "更新文章",
        "description": "根据文章ID更新文章信息",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "文章唯一标识符"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "文章标题"
                  },
                  "content": {
                    "type": "string",
                    "description": "文章内容"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "文章标签ID列表"
                  },
                  "category": {
                    "type": "string",
                    "description": "文章分类ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "文章更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "article": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "没有权限更新文章"
          },
          "404": {
            "description": "文章未找到"
          },
          "500": {
            "description": "更新文章失败"
          }
        }
      },
      "delete": {
        "summary": "删除文章",
        "description": "根据文章ID删除文章",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "文章唯一标识符"
          }
        ],
        "responses": {
          "200": {
            "description": "文章删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "没有权限删除文章"
          },
          "404": {
            "description": "文章未找到"
          },
          "500": {
            "description": "删除文章失败"
          }
        }
      }
    },
...
    "/blog/categories": {
      "get": {
        "summary": "获取分类列表",
        "description": "分页获取所有文章分类",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取分类列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "categories": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "获取分类列表失败"
          }
        }
      },
      "post": {
        "summary": "创建新分类",
        "description": "创建一个新的文章分类",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "分类名称"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "分类创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "category": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "创建分类失败"
          }
        }
      }
    },
    "/blog/categories/{id}": {
      "put": {
        "summary": "更新分类",
        "description": "根据分类ID更新分类信息",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "分类唯一标识符"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "分类名称"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "分类更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "category": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "分类未找到"
          },
          "500": {
            "description": "更新分类失败"
          }
        }
      },
      "delete": {
        "summary": "删除分类",
        "description": "根据分类ID删除分类",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "分类唯一标识符"
          }
        ],
        "responses": {
          "200": {
            "description": "分类删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "分类未找到"
          },
          "500": {
            "description": "删除分类失败"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}